Create RoomsSetBehaviourProtection out of redundant protections.
